# openapi-v1.yaml
...
openapi: 3.1.0
info:
  title: Nethical Governance API v1
  description: |
    Enhanced REST API for AI safety and ethics governance with full backend management.
    
    ## New Features in v1
    - **RBAC**: Role-based access control with JWT authentication
      - `admin`: Full access to all operations
      - `auditor`: Read-only access to logs and audit data
      - `operator`: Can evaluate risk, but cannot modify configuration
    - **Agent Management**: CRUD operations for AI agent configuration
    - **Policy Management**: CRUD operations for governance policies
    - **Audit Logs**: Read-only access with Merkle tree verification
    - **Real-time Threats**: WebSocket and SSE for live threat notifications
    
    ## Authentication
    All v1 endpoints (except /auth/login) require authentication via Bearer token.
    
    ### Login Example:
    ```bash
    curl -X POST http://localhost:8000/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"username": "admin", "password": "admin123"}'
    ```
    
    ### Using Token:
    ```bash
    curl http://localhost:8000/api/v1/agents \
      -H "Authorization: Bearer YOUR_TOKEN_HERE"
    ```
    
  version: 1.0.0
  contact:
    name: Nethical Team
    url: https://github.com/V1B3hR/nethical
    email: support@nethical.dev
  license:
    name: MIT
    url: https://github.com/V1B3hR/nethical/blob/main/LICENSE
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
...
paths:
...
  /hub/dock:
    post:
      tags:
        - Nethical Hub & Dock
      summary: Dock an agent
      description: |
        Securely dock an AI agent spaceship into the Nethical Hub.
      operationId: dockAgent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agent_id
              properties:
                agent_id:
                  type: string
      responses:
        '200':
          description: Agent docked successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Docking rejected by governance
        '404':
          description: Agent not registered
  /hub/undock:
    post:
      tags:
        - Nethical Hub & Dock
      summary: Undock an agent
      description: |
        Undock an AI agent from the Nethical Hub.
      operationId: undockAgent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agent_id
              properties:
                agent_id:
                  type: string
      responses:
        '200':
          description: Agent undocked successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /hub/active:
    get:
      tags:
        - Nethical Hub & Dock
      summary: List active docked agents
      description: |
        Retrieve all currently docked and visible AI spaceships.
      operationId: listActiveHubAgents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully retrieved active agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /hub/exchange:
    post:
      tags:
        - Nethical Hub & Dock
      summary: Exchange message between agents
      description: |
        Transmit a secure, audited message between two docked AI spaceships.
      operationId: exchangeHubMessage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HubMessage'
      responses:
        '200':
          description: Message delivered successfully
        '400':
          description: Threat detected by port gateway
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Transmission blocked by trust policy
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    AgentResponse:
      type: object
      properties:
        id:
          type: integer
        agent_id:
          type: string
        name:
          type: string
        agent_type:
          type: string
        description:
          type: string
        trust_level:
          type: number
        status:
          type: string
        configuration:
          type: object
        metadata:
          type: object
        dock_status:
          type: string
        visibility:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
    HubMessage:
      type: object
      required:
        - sender_agent_id
        - recipient_agent_id
        - payload
      properties:
        message_id:
          type: string
        sender_agent_id:
          type: string
        recipient_agent_id:
          type: string
        intent:
          type: string
        payload_type:
          type: string
        payload:
          type: string
        ttl:
          type: integer
        trust_required_level:
          type: number
        timestamp:
          type: string
          format: date-time
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...